Update easyeffects icon (#236)
* Update easyeffects icon

They designed a new icon back in 2022. It still fits nicely with this theme so let's update.

More importantly, the current icon misrenders on KDE Plasma because the QT SVG renderer doesn't
support <clipPath>. The new icon doesn't make use of that feature and renders fine.

* Add symbolic easyeffects icon

Fits much better with the rest of the symbolic icons in the tray.