Fix loading new KeepassXC main icon
KeepassXC looks for its icon under "keepassxc.svg", but the new icon version was added under the
"keepassxc-unlocked.svg" name. Since "keepassxc.svg" was still a symlink to "keepass.svg", the old
icon still ended up being loaded.

- Delete the "keepassxc.svg" symlink.
- Rename "keepassxc-unlocked.svg" to "keepassxc.svg".
- Create a new symlink "keepassxc-unlocked.svg" pointing to "keepassxc.svg".

It is unclear when or if "keepassxc-unlocked.svg" ever gets requested, but the symlink is left to
be on the safe side.